Extract LabelPropertiesIndicesInfo to lightweight header
Removes heavy label_property_index.hpp include from index_lookup.hpp
by extracting LabelPropertiesIndicesInfo to its own header with only
lightweight dependencies (id_types, index_order, property_path).
